+++ Scripts/Camera/CameraTarget.cs
@@ -33,16 +33,60 @@ namespace IBM.Watson.DeveloperCloud.Camera
         private static UnityEngine.Camera mp_CameraAttached = null;
         [SerializeField]
         private bool m_UseCustomPosition = false;
+        [SerializeField]
         private Vector3 m_CustomPosition = Vector3.zero;
+        [SerializeField]
+        private Vector3 m_OffsetPosition = Vector3.zero;
+        private Quaternion m_OffsetPositionRotation = Quaternion.identity;
         [SerializeField]
         private bool m_UseCustomRotation = false;
         private Quaternion m_CustomRotation = Quaternion.identity;
         private bool m_UseTargetObjectToRotate = false;
+        [SerializeField]
         private GameObject m_CustomTargetObjectToLookAt = null;
+
+        [SerializeField]
+        private bool m_TextEnableCamera = false;
+        [SerializeField]
+        private bool m_TestToMakeItCurrent = false;
 
         #endregion
 
         #region Public Members
+
+        /// <summary>
+        /// Gets or sets a value indicating whether this <see cref="IBM.Watson.DeveloperCloud.Camera.CameraTarget"/> use
+        /// custom position.
+        /// </summary>
+        /// <value><c>true</c> if use custom position; otherwise, <c>false</c>.</value>
+        public bool UseCustomPosition
+        {
+            get
+            {
+                return m_UseCustomPosition;
+            }
+            set
+            {
+                m_UseCustomPosition = value;
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether this <see cref="IBM.Watson.DeveloperCloud.Camera.CameraTarget"/> use
+        /// custom rotation.
+        /// </summary>
+        /// <value><c>true</c> if use custom rotation; otherwise, <c>false</c>.</value>
+        public bool UseCustomRotation
+        {
+            get
+            {
+                return m_UseCustomRotation;
+            }
+            set
+            {
+                m_UseCustomRotation = value;
+            }
+        }
 
         /// <summary>
         /// Gets or sets the target position.
@@ -57,9 +101,13 @@ namespace IBM.Watson.DeveloperCloud.Camera
                 {
                     return m_CustomPosition;
                 }
+                else if (m_OffsetPosition != Vector3.zero)
+                {
+                    return transform.position + ( Quaternion.Euler(transform.rotation.eulerAngles - m_OffsetPositionRotation.eulerAngles) * m_OffsetPosition);
+                }
                 else
                 {
-                    return transform.position;
+                    return transform.position ;
                 }
             }
             set
@@ -77,11 +125,7 @@ namespace IBM.Watson.DeveloperCloud.Camera
         {
             get
             {
-                if (m_UseCustomRotation)
-                {
-                    return m_CustomRotation;
-                }
-                else if (m_UseTargetObjectToRotate)
+                if (m_UseTargetObjectToRotate)
                 {
                     if (TargetObject != null)
                     {
@@ -102,6 +146,10 @@ namespace IBM.Watson.DeveloperCloud.Camera
                         return Quaternion.identity;
                     }
                 }
+                else if (m_UseCustomRotation)
+                {
+                    return m_CustomRotation;
+                }
                 else
                 {
                     return transform.rotation;
@@ -114,7 +162,11 @@ namespace IBM.Watson.DeveloperCloud.Camera
             }
         }
 
-        protected GameObject TargetObject
+        /// <summary>
+        /// Gets or sets the target object.
+        /// </summary>
+        /// <value>The target object.</value>
+        public GameObject TargetObject
         {
             get
             {
@@ -122,12 +174,24 @@ namespace IBM.Watson.DeveloperCloud.Camera
             }
             set
             {
-                m_UseTargetObjectToRotate = true;
-                m_CustomTargetObjectToLookAt = value;
+                if (value != null)
+                {
+                    m_UseTargetObjectToRotate = true;
+                    m_CustomTargetObjectToLookAt = value;
+                }
+                else
+                {
+                    m_UseTargetObjectToRotate = false;
+                    m_CustomTargetObjectToLookAt = null;
+                }
             }
         }
 
-        protected UnityEngine.Camera CameraAttached
+        /// <summary>
+        /// Gets the camera attached.
+        /// </summary>
+        /// <value>The camera attached.</value>
+        public UnityEngine.Camera CameraAttached
         {
             get
             {
@@ -140,7 +204,11 @@ namespace IBM.Watson.DeveloperCloud.Camera
             }
         }
 
-        protected WatsonCamera WatsonCameraAttached
+        /// <summary>
+        /// Gets the watson camera attached.
+        /// </summary>
+        /// <value>The watson camera attached.</value>
+        public WatsonCamera WatsonCameraAttached
         {
             get
             {
@@ -156,7 +224,11 @@ namespace IBM.Watson.DeveloperCloud.Camera
 
         #region Set Target on Camera
 
-        protected void SetCurrentTargetOnCamera(bool enable)
+        /// <summary>
+        /// Sets the current target on camera.
+        /// </summary>
+        /// <param name="enable">If set to <c>true</c> enable.</param>
+        public void SetCurrentTargetOnCamera(bool enable)
         {
             if (WatsonCamera.Instance != null)
             {
@@ -167,7 +239,10 @@ namespace IBM.Watson.DeveloperCloud.Camera
             }
         }
 
-        protected void SetTargetPositionDefault()
+        /// <summary>
+        /// Sets the target position default.
+        /// </summary>
+        public void SetTargetPositionDefault()
         {
             if (WatsonCamera.Instance != null && WatsonCamera.Instance.DefaultCameraTarget != null)
             {
@@ -175,12 +250,38 @@ namespace IBM.Watson.DeveloperCloud.Camera
             }
         }
 
-        protected void SetTargetRotationDefault()
+        /// <summary>
+        /// Sets the target rotation default.
+        /// </summary>
+        public void SetTargetRotationDefault()
         {
             if (WatsonCamera.Instance != null && WatsonCamera.Instance.DefaultCameraTarget != null)
             {
                 TargetRotation = WatsonCamera.Instance.DefaultCameraTarget.TargetRotation;
             }
+        }
+
+        #endregion
+
+        #region Update
+
+        void Update()
+        {
+            if (m_TestToMakeItCurrent)
+            {
+                m_TestToMakeItCurrent = false;
+                SetCurrentTargetOnCamera(m_TextEnableCamera);
+            }
+        }
+
+        #endregion
+
+        #region public Functions
+
+        public void SetTargetPositionWithOffset(Vector3 offsetPosition)
+        {
+            m_OffsetPosition = offsetPosition;
+            m_OffsetPositionRotation = this.transform.rotation;
         }
 
         #endregion